"4:10pm. I am up. Yeah, I am trough with trying to figure things out myself. It is only games for me from here on out. The brainless nature could do it, and me despite having intelligence can't. This is wrong. I am wrong.
My approach is definitely wrong. It should never be the case that adding intelligence to a brainless process should make it worse.

If it is wrong, then the setup itself is wrong.

Where I went wrong is thinking I had to figure out intelligence.

Rather, back in high school I should have designated nature as an enemy and seriously fought it. And the best way to fight it is to immitate it.

The self improvement game has to be an evolution game.

Since I am getting into game dev, that is the kind of game I should be fundamentally making. Take poker, give it an evolutionary manifold, and see what algorithms end up being dominant. Study those. That is how you get true insight.

Instead I am just bashing my head against the wall. I've been trying to reach a height that is unatainable to me. You can't get insight purely from imagination.

I do not care if people are having a wild party and making money with deep learning. This is not what I signed up for. To nurture my interest, I want a steady stream of good algorithms.

4:20pm. Maybe it is finally time to take a trek back to 2015. Back then I was looking up genetic programming, but did not go far with it because I got interested in deep learning. John Koza wrote books on that, and I should be able to find them on Libgen.

Though I do not have the hardware for it. And I do not want to endure the loneliness of yet more failed ventures. I just can't afford to keep doing upaid work forever. If I could get 1k per month to work on my own stuff things would be completely different.

https://www.reddit.com/r/GroqInc/

It could be worth a try to make a post here, but I doubt it would work. This yearning is just so painful. Having a product not in demand is the worst thing ever.

It is all the ML community's fault. If they gave me algorithms that can beat poker cleanly I would not be in this mess.

4:30pm. Well, let me try it.

https://www.anandtech.com/tag/tenstorrent

Oh lol, I almost posted it on the wrong Reddit. I was wondering why I could not find the AnandTech interview link.

Though I should post it on the Groq one as well if this fails.

https://www.reddit.com/r/Tenstorrent/

///

As the 20s proceed, novel AI hardware to replace GPUs will arrive at the scene, but by the looks of things Python + C++ seems like it will be a dominant combination for programming them. This is a huge pity as we could do better than languages created in the 80s and 70s for programming hardware created in the 2020s. I am trying to change this destiny with [Spiral](https://github.com/mrakgr/The-Spiral-Language).

Apart from some special features to control inlining and specialization which make it suitable for hardware with no dynamic memory allocation capability like the GPUs, Spiral is quite similar to languages like F# and OCaml. It has static typing, global type inference, first class functions, records, tuples, unions and more as any competent functional language would. Unlike toy languages, it has a well done language server - if you want to try out the language, it is as simple as installing its plugin in the VS Code marketplace. About 3 years of full time work went into it.

I read the [Anandtech interview](https://www.anandtech.com/show/16709/an-interview-with-tenstorrent-ceo-ljubisa-bajic-and-cto-jim-keller) by the CEO and CTO of TensTorrent who expressed their desire to support general purpose programming on their chips. I'd like to support that. If you are sincere in that desire you should consider giving me a [grant](https://opencollective.com/spiral-collective) and access to your hardware so I can create a ref counted C backend for Spiral.

This should make them a lot easier to program in than raw C, I can promise you that.

I myself am personally interested in using the hardware and would do some evolutionary RL stuff and game dev which would serve as a demo.

Background: See my [resume](https://docs.google.com/document/d/e/2PACX-1vSe88lhaoSZaF-YHKeiA_RlpkPwzLetrIAoY2v6ckbaLn59QQ95nvMC-Cc13hxDsLzVJ-KGE824S1bF/pub).

Would you be interested in supporting language development for your hardware?

///

Let me go with this. Probably some members of the company should see it and think about it. I tried emailing the CEO directly months ago, but that did not work. And I am certainly not going to go through HR. For such an offer, doing it on their social media page would at least get them to think about it."