"https://www.reddit.com/r/Tenstorrent/comments/rhtuxn/seeking_a_sponsor_for_a_functional_programming/
Here it is. Now instead of dwelling in my own mind let me finally watch some videos. I am going to try TensTorrent and Groq, and maybe Graphcore, but if that fails, there really is no point in putting any more investment into Spiral. It will have to stay dead for a few years while I do other things. I should have done this a few months ago, as it would be the ideal way of getting feedback. Instead I wasted my time trying to go through regular channels.

Let me watch a video by Grant.

I need to lift my spirit.

https://www.blenderkit.com/articles/introducing-texture-resolutions-for-models-materials-and-hdrs/

Where do I find the textures for BlenderKit?

https://www.youtube.com/watch?v=7ax_x_L5AmI
How to use Blender Kit Online Asset Library Addon in Blender 2.8 Series - Addon Tutorial Series

Let me just watch this again.

https://youtu.be/7ax_x_L5AmI?t=94

Ah, right. I need to select the material. But that just gives me the wood node instead of a texture. I am not sure how to deal with that.

5:30pm. Ok, nevermind. I get it. While the materials are good, they aren't what I need here. I need actual textures.

https://ambientcg.com/view?id=Bricks075A

Let me go for bricks. Oh wow, the 16k version is 1.7Gb. I'll go for the smallest one.

5:45pm. I do not know why, but seams from islands refuses to work for me.

Ah, I get it. I need to select it all on the UV side as well. Right.

6:10pm. I have no idea. The normal map simply refuses to work. I'll have to figure that out later.

6:15pm. Focus me, what is next?

https://youtu.be/OFxeVQOZ8nY
Nodes 4 Noobs | Lvl 5 | PBR MATERIALS | Beginners Guide to Nodes

How about this?

https://www.youtube.com/watch?v=4H5W6C_Mbck
How to make Photorealistic Materials in Blender, using the Principled Shader

He recommends that I watch a video on PBRs first. So let me do it. It is 38m long.

https://youtu.be/4H5W6C_Mbck?t=388

Ah, wait. If the principled shader replaces the other ones, what about the transparent shader? Should it not replace it as well? It should be possible, but I haven't managed to get the alpha settings to do anything. It might be worth asking on SO later.

https://youtu.be/4H5W6C_Mbck?t=972

This is interesting. I suppose it would be good for skin.

7:05pm. Done with lunch. Let me watch the video to completion.

https://youtu.be/4H5W6C_Mbck?t=1219

This is a very informative video on the principal shader. It really was the right choice to watch it.

7:30pm. https://youtu.be/4H5W6C_Mbck?t=2151

Ahh, so this is how normal maps should be used.

7:30pm. It works, it works. It looks quite nice when seen from a distance. Though now I am quite interested in how displacement maps work.

7:40pm. Ok, I think I get what clearcoat is. It does not seem like it has a color related to it. Some of these effects are really subtle. Let me go back to the video by Grant.

7:45pm. The idea of using Reddit to seek sponsors like this is a really good one, though my emotional reaction to it is numbed. I should have done this as my very first move. Lack of experience leads to bad decisions. But I would not have thought of it to begin with had I not stepped away from the screen to just think.

https://youtu.be/OFxeVQOZ8nY?t=147

Ah, so this is how you plug in a HDRI. I'll remember it this time.

https://youtu.be/OFxeVQOZ8nY?t=222

I had no idea you could just drag the files like this directly. This is quite nice.

7:55pm. https://youtu.be/OFxeVQOZ8nY?t=407

Why doesn't he just use a sphere projection? Also why does the UV map do that by default?

8pm. https://youtu.be/al3vR6ySXfo
Nodes 4 Noobs | Lvl 6 | Displacement | Beginners Guide to Nodes

I'll leave this for tomorrow. Right now, I am in a slump, but I will work through it. I can only blame my own lack of skill for having to rely on the whims of fate to give me my opportunity.

Life is long, and even if I spend a couple of things doing creative work, it is not like programming skills will disappear. I haven't riden a bike in two decades, but I have confidence that I'd be able to do it regardless.

It is not like I do not want to work, but I definitely do not want to work on things outside my interest zone. Long projects like a PL require transparency around one's intentions. If AI chips go the C route despite my wishes, the benefits of Spiral will remain just for myself, but that is a benefit on its own.

Getting good at art will be difficult, but what are few years here and there? Compared to the last 7, I'll master this in a flash.

Right, before I close...

https://blender.stackexchange.com/questions/90392/how-do-i-get-transparency-to-work-with-the-principled-bsdf-shader

Exactly what I wanted to ask.

...This is an interesting way of doing it, but just what is the principled alpha doing then?

https://blender.stackexchange.com/questions/46424/add-a-transparent-image-on-top-of-a-material/

I do not have the mental energy to parse all this right now. I'll do more research into the question tomorrow. I feel like transparency should be achievable in just the principled shader, but I do not know how.

Assuming I can focus, tomorrow I should be able to the rest of the series and get back to the piece I am working on. That should be more fun than just watching tuts for hours on end."