"10:20am. I am up. I wonder if I'll get replies to the TensTorrent query. At any rate, posting it there is the best idea ever as I'll be able to decisively get rid of them not knowing about the language scenario. If they don't want it then so be it. I'll be able to move forward knowing that this is the case, and not have to be concerned about not doing enough to get it noticed.
No, nothing yet. The sub is very inactive, so I'll have to wait for a while until somebody from the company sees it. I'll post it on Groq and Graphcore's pages as well. It should happen eventually. At some point the CEO or somebody higher up will log into their social media page and see my post.

10:30am. Let me chill for a bit and thne I will start. The extreme tension I've been feeling for the past few days has abbated. This is the correct strategy for seeking sponsors. Doing just this while I do my own things on the side is enough.

I was wrong to ever try applying at places. Applying to jobs should be done for the sake of actually getting a job.

11:10am. Let me start. Where was I yesterday?

https://youtu.be/al3vR6ySXfo?t=39
Nodes 4 Noobs | Lvl 6 | Displacement | Beginners Guide to Nodes

This looks really great.

11:25am. https://youtu.be/al3vR6ySXfo?t=356

Let me actually try this.

11:30am. Hmmm, how do I turn on displacement? I fact, I should not forget to do a subsurface mod first. ...But that is complicated.

11:35am. I figured out how to get it to work, but it seems it only works in Cycles. Still it looks quite nice there. I wonder why suddenly it is wanting to get 1024 samples whereas previously 32 was max. I don't think I changed any parameter by accident. Edit: I forgot that I upgrade to Blender 3 which has rendering performance improvements. This is probably related to that.

https://youtu.be/al3vR6ySXfo?t=405

Hmmm, there is a modifier for this.

12:15pm. https://youtu.be/oYtfY1ESMwU
Nodes 4 Noobs | Lvl 8 | Pointiness & Blend Modes | Beginners Guide to Nodes

I think this should be the last one. Let me check.

https://www.youtube.com/playlist?list=PLn3ukorJv4vtnU_TaZob7QD6Q8d9C9Ki7

Ah, there is also the mossy rocks exercise in the playlist.

I'll take a look at it after I am done with this one.

12:35pm. https://youtu.be/oYtfY1ESMwU?t=233

I have no idea what this is supposed to be doing. This is one of those cases where I really need the guy to explain it to me. Plenty of such cases in Blender. Blender is so complex that even the stuff that is easy is hidden away amongst piles of other options. A lot of stuff needs to go into my procedural memory to make all of this work.

12:40pm. Ohhh, ok I see it. For making this work it is best to edit the pos of the slider nodes manually. When I set it to .495/.505 I really see the crevices getting darker. Quite an interesting effect. I'll have to remember this trick for the future.

https://youtu.be/oYtfY1ESMwU?t=361

Oh, this is an interesting way of doing it. Yes, this is one way to get precision.

1:30pm. Let me have breakfast here. I got into it. Today's session was ideal. I learned a bunch.

My heart is finally at peace now. It is all because I could imagine a future where these chips get sold and have customers congregating on Reddit. That is the place I could advertize. If my current efforts fail, I could implement that backend, push tutorials and such. No prob.

Right now, these chips are still in early stages, so it does not matter that I start work immediately. But it does matter that I start advertizing Spiral immediately.

I'll finish this tutorial and then post on the Groq and Graphcore pages. Maybe I should look into SambaNova as well."