"9:50am. The internet is down. Don't tell me the budget is spent already?
I actually must have woken up an hour or two ago and just spent my time in thought.

I've decided - if Spiral gets no bite, and in the future I am forced to finance all my research out of my own pocket, I am going to leave the public work done here as it is and take the genetic programming work, the AI backends and work on the language private.

Hmmmm, it is still not back. Let me change it to the slow line.

10:25am. Ok, I am almost ready to start. Let me chill a bit more and then I'll get on with it.

10:50am. Let me start. It is time to get busy with the HDRI. I really needed a break after yesterday. I finished at 9pm after all. Let me not rush myself and do it one step at a time. I had a good long morning break.

I still need some time to settle my emotions, but it will happen over time.

I am thinking what I am going to do that AI chip, and something like doing a ref counted interpreter for the sake of GP research would be a PhD difficulty project for the average researcher. I myself could do it in a month or two, but imagine anyone trying to do this in C or even Assembly. It would be very difficult.

This is providence. My programming skills couldn't get me to my goal directly, but I'll get a reward for my language work eventually. Spiral could be a good for a lot of things, but making the AI chip into an oracle that will give me the true ML algorithms is something it is uniquely suited for. I am going to get divine feel of inspiration like in 2013 when I first came to an understanding of the self improvement loop once again.

But today, I will prove my will through art.

11:05am. Maybe I will be done with the HDRI. Maybe tomorrow. And maybe the day after. It does not matter. The time is going to get invested in it.

First of all, I need the basics. Let me make a jewel, a simple 6 vertex object. Then I will think about how distribut it around a shell of some other object.

This is simple, but I do not have any experience with geometry nodes. I've played with shaders so I can find my way around them now, but I've only watched tutorials for geometry nodes.

11:15am. I just about managed to remember that it is possible to drag the object from the collection window into the node window. That saves me doing the object info node.

11:20am. Oh, wow, this actually went a lot smoother than it would.

I used the Instance on Points node, connected the Group Input to Points, and the shard to the instance. And I passed that the output and as a result I have the shard on all the sphere points.

11:35am. I am trying to get my bearings. Geometry nodes have textures. They can even distribute volume to points. This stuff is powerful.

11:40am. How do I randomize the points?

Let me take a short break here.

12pm. Let me resume, I had to run. I should take a longer break here, but I want to grind at it for a while longer. Right now I am not doing much, and I am seemingly wasting time, but this kind of practice is important to internalize the material.

12:05pm. I had some time to think. If the problem was just to randomize the positions, I could use subdivide on the original mesh and then use the fractal option to make it noisy.

But I also want to randomize the rotations. I can only do that in geometry nodes. That having said, the random value option is just for scalars. it has the ID option which I am guessing could be used to generate a bunch of them, but I do not know how combine the scalar socket with geometry socket that vectors use.

I do not know. I am going to have to watch some tuts on this. How do I add noise to instance positions and rotations? That is the question. Let me start by searching for searching for...

https://www.youtube.com/results?search_query=geometry+node+random

https://www.youtube.com/playlist?list=PLgO2ChD7acqHzccBuhAGw8dTPLnR1E3QB

This is a whole series on just geometry nodes. Given what I am working on, this might be a good idea to go through. Now that I have an actual use case in mind, I'd be really motivated to learn this.

https://www.youtube.com/watch?v=KntGPD1k7v0
07-Points and Instancing -Geometry Nodes( Fields) For Beginners / Blender 3.0

Let me start near the end since that is what Google gave me.

https://youtu.be/KntGPD1k7v0?t=185

He already went beyond what I could think of. Those speed tutorials are one thing, but these in depth courses are really valuable. Programming is one thing, but Blender really goes well with video for obvious reasons. These Youtube videos are a godsend. Imagine trying to learn Blender without them, I'd never get anywhere.

https://youtu.be/KntGPD1k7v0?t=302

Ahhh, right. I could use the noise texture to randomize the points. Why didn't I think of that?

He is working really fast here. Instead of adding the position to the Musgrave texture, he could have plugged it in directly into it.

https://youtu.be/KntGPD1k7v0?t=374

I have everything I need here. Let me give it a try.

12:50pm. Oh, I got a good effect. I quite like this. Adding the noise texture to the position offset is the way to go. Since 3d noise texture duplicates are equivalent I used the 4d one for rotation.

1pm. I've parameterized the size as well using the map range node.

1:05pm. If I am to describe it, right now it looks like an asteroid field wrapped in a sphere. This is quite good.

1:10pm. I am an easy person to please. I am just stading here in admiration of my own work. This was a really simple task, and I've conquered it. The larger shards are quite dominant, but not overbearing. I'll want to add more spheres. I need more smaller shards to populate the scenery. Maybe I'll add a few big ones by hand.

1:15pm. Let me stop here. It is time for breakfast. In honor of finishing this first step, I should watch the other geometry nodes videos from the course.

Though before that, I should just finish this project. It won't be hard at all. And going through course after that will allow me to reconsider my choices in light of new knowledge afterwards. That will be the best way to learn the material."