"https://www.reddit.com/r/hardware/comments/mx3hvl/jim_kellerled_tenstorrent_licenses_riscv_for_ai/
> Offhand, their newly released Grayskull Soc/NoC is only pulling about 4.9TOPS/W, which is better than Intel or Nvidia, but about half the performance of bleeding edge Process In Memory architectures that will enter the market soon. Different approaches will scale better than others though, and Tenstorrent may have an edge there.

I wonder which PIM architectures there are?

12/25/2021

10:25am. It is Chrismas already. Time passes too fast for me, it feels like it was the 21st yesterday. Less that a week till the yearly review.

Let me chill a bit and I will start. That /r/hardware thread and...

https://spectrum.ieee.org/samsung-ai-memory-chips

Indicate well enough that there will be plenty of PIM chips to choose from in the future. I do not need to be obsessed about Grayskull.

11:10am. Let me start. Yesterday once I finished for the day, I was a degenerate, just spending my time lurking 4chan threads and reading manga. Right now I am doing that too. In the background I keep bolstering my desire to do art

I've made a decision, it is really difficult to have a 6-vertex shard have the kind of shape that I'd want it to. Instead I should bevel it and make it more spherical. That will make it a lot easier to use facing to give it a golden outline. It will really increase the computational complexity of the scene, but that is fine.

11:15am. Today let me first focus on the geometry node tuts. Then I will properly dive into it and deal with the HDRI.

https://www.youtube.com/watch?v=L_8xTV3IP3A&list=PLgO2ChD7acqHzccBuhAGw8dTPLnR1E3QB&index=5
04 Applying Materials -Geometry Nodes For Beginners / Blender 3.0

Let me start this.

It is time to get some learning done. After that I'll make things happen.

https://youtu.be/L_8xTV3IP3A?list=PLgO2ChD7acqHzccBuhAGw8dTPLnR1E3QB&t=68

Ah, right. This made me remember that you can assign different material to each face. But in that case, how is volume associated with faces?

https://youtu.be/L_8xTV3IP3A?list=PLgO2ChD7acqHzccBuhAGw8dTPLnR1E3QB&t=621

Hopefully they will have fixed this by now.

https://youtu.be/L_8xTV3IP3A?list=PLgO2ChD7acqHzccBuhAGw8dTPLnR1E3QB&t=746

I should follow the next video step by step instead of trying to save time.

11:45am. https://youtu.be/_PWaBW5uJfE?list=PLgO2ChD7acqHzccBuhAGw8dTPLnR1E3QB&t=660

Damn, the project files cost 12$. How sneaky of him! Nevermind following this right now. I just need to get the gist of it. For a programmer like me, this is not difficult.

https://youtu.be/_PWaBW5uJfE?list=PLgO2ChD7acqHzccBuhAGw8dTPLnR1E3QB&t=899

Instead of doing it like this it would have been easier to use a group.

https://youtu.be/_PWaBW5uJfE?list=PLgO2ChD7acqHzccBuhAGw8dTPLnR1E3QB&t=1060

This is interesting. I started wondering how I would do a snap as soon as he mentioned it. To think this exists.

https://youtu.be/_PWaBW5uJfE?list=PLgO2ChD7acqHzccBuhAGw8dTPLnR1E3QB&t=1256

These kinds of landscapes. I want to see how that is done.

https://youtu.be/RhiJQlwD98A?list=PLgO2ChD7acqHzccBuhAGw8dTPLnR1E3QB&t=561

Rather than multiply it would be better to use separate textures for each direction. Or rather, use it as a displacement map. Then it would work on a sphere.

https://youtu.be/RhiJQlwD98A?list=PLgO2ChD7acqHzccBuhAGw8dTPLnR1E3QB&t=678

This looks quite good. So that is how he got it.

https://youtu.be/RhiJQlwD98A?list=PLgO2ChD7acqHzccBuhAGw8dTPLnR1E3QB&t=746

Using a sky texture for the background. That is something I did not think of.

12:25pm. https://youtu.be/KntGPD1k7v0?list=PLgO2ChD7acqHzccBuhAGw8dTPLnR1E3QB
07-Points and Instancing -Geometry Nodes( Fields) For Beginners / Blender 3.0

Finally back to where I was yesterday. Now I can watch this. So far these videos have been easy to grasp and I am not loosing much by not following the individual steps. I can understand them well enough. As far as programming complexity goes, this is quite light.

Let me stop here for breakfast.

After that I'll clear the playlist and deal with the HDRI. I've been messing it for too long. One thing that is bothering me right now, is that I've messed up the the limbo. I joined some vertices with f instead of j, and I think I have tears. I am going to have to redo it unfortunately, but that should not be too hard.

https://youtu.be/N8O0bE898Qc
Get Good At Blender 3 - The Shear Tool

Let me just watch this.

https://youtu.be/N8O0bE898Qc?t=135

Hmmmm...this might be a better way of doing it that I would have done. I'd have put in a loop cut before extruding, but that would have the disadvantage of adding unecessary geometry. This way is ideal.

12:45pm. I really hate the way I extruded the walkway. I should have left the original as a sphere and made the walkway a separate object. Or better, made it a curve and proceduraly generate the walkway. I need to up my skills.

I am going to redo it and make better use of booleans. Now that I have some exp, it should not take me hours to do like last time.

Let me get breakfast. Focus."