"10:50am. Done with chores. I had trouble falling asleep due to change in temperature, and I ended up inadvertently being waken up early by my parents. Let me chill a bit and then I will start.
11:25am. Let me start. Time to get those stars done.

11:25am. Instead of doing it like I have before, let me try the magic wand, though I doubt that will work.

11:40am. It did not work all at once, but using the auto select was a lot easier than using the g-pen. Now I have my stars. Let me give them the golden outlines.

12:25pm. Did the golden outlines for the stars. It makes them look quite nice. I could take what I have now and use it as a wallpaper.

12:30pm. This is a great spot to stop for breakfast. I'll work on reflections after breakfast.

1:25pm. Just a bit more and I will resume.

1:55pm. Let me resume. It is time to deal with the painting.

3:15pm. Right now I am on very last part. Just a bit more and I am done.

3:50pm. I am done. I did it!

Hell yeah.

4pm. I guess I'll post it in the /beg/ thread. Since I have my first real artwork, I should look into opening an artist's account somewhere. I have no idea where.

5pm. Where did the last hour go? I feel high right now. I've even sent this to my dad's email to impress him. At least one person in the /beg/ thread likes it. I am proud.

Ok, I'll set it as my wallpaper even though it would waste memory.

5:10pm. The taskbar cuts off the lower part, what a pity. But it is fine.

5:15pm. Ok, focus me. I'll want to do some research on where to post my art online. But otherwise it is yearly review time.

5:20pm. Focus me, focus.

https://share.clip-studio.com/en-us

Something like this? I don't know.

Tumblr? I have no idea. Though I won't make porn, I do not want the platform getting in the way of me posting naked girls.

Ahh, what I need is something like Pixiv.

5:35pm. Enough of this, I'll just ask in the /beg/ thread. I'll look up advice on Youtube on where to post art if that fails. Youtube hasn't failed me so far when it comes to education.

5:40pm. I need to focus on the review. Right now my brain is euphoric from finishing the piece and getting praised a bit. I want to just call it a day here, but let me at least start the review.

///

In the last review, I was in a middle of a heated battle. I though with just a push I'd be able to make the other side crumble. Bullets were blazing, but expected I would be able to push through because surely something like training a poker should be doable given how near the Singularity is. What happened that in the frenzy of battle I lost track of ammo and when I went to reload, I found my satchel empty. And the enemy which was on the verge of defeat got a fresh batch of reinforcements, ready to open fire.

I suffered a lot of mental damage when I threw in the towel back in September, and just about now my sanity points have recevered to a reasonable level. To think that in the last post I thought I was only a few weeks away from getting the agent to work.

There is no getting the poker agent plan to work with my current level of hardware. None.

The methods I had worked just fine on toy games like Leduc and even Flop poker, but pretty much everything I've tried just dies on Holdem. The only thing that worked for me was increasing the batch size by 10x to 5k, but that slows down the already very slow training to an unbearable degree. The realization that I do not have the computational power to fulfil my goal made me extremely obsessed about AI chips, to the point I actually considered getting a job for the first time, just to get the chip. During my bouts of normality, I did spent time making a resume and applying. One time I even got an offer, but it was so poor that I gave up in disgust. Maybe I shouldn't as it would have been enough to get the chip, but I honestly felt the other party was mocking me with how the negotiation went. I do not regret aborting it. Plenty of places list salary ranges and I should have stuck to those when applying.

The tech job market is such a shitshow, the reject rates make it impossible to pick something you'd like to work on, instead if you are really serious, you have to pick whatever you get. You can either focus on maximizing your salary or picking work that is meaningful.

To continue on the path of developing my external cortex I need the bare minimum of money to buy one of the Grayskull chips from TensTorrent. They cost 1-2k might not be much depending on where you live, but I have no way of getting that amount without doing paid labor. This pisses me off because I was supposed to make money of poker to make those chips in the first place. For the first time, I've felt that the path I am on is particularly weak. It is just sending me hurdles, but it is not sustaining me wit resources to keep pursuing it. It was one thing when my goal was just to make a language and a ML library, but now I need the world to cooperate.

This situation made me reflect upon my approach, and got me thinking. It is one thing try to get to human or even animal level, but I was so sure that CPU+GPU should be enough for toy games, which I'd say poker falls into. I did not expect it to get to superhuman level even there, but I wasn't prepared for the amount of struggle the crappy current day algorithms would have to endure. At this point I can only ask: if CPU+GPU aren't enough, just how sure am I that getting an AI chip would help? I mean, by porting the game directly to chip and parallelizing the training, I could get 100x performance improvements most likely. That would be enough to cover the increases in batch size necessary to do training. That is obviously right.

But what then? Past poker I'd need to scale again and run into the same again. A single chip is not going to cut it on Dota or Starcraft. Unlike Deepmind even if I had the money, I can't just open my wallet every time I run into problem. I actually want to make money off RL, not go deeply into the red for sake of research.

At the start of the year, I forcefully quashed my skepticism towards deep learning and backprop, but I think my initial impression was right after all.

If I am going to succeed, I need better algorithms. Backprop was the only choice I had, so I had to go for it, but that was a mistake.

///

7pm. Done with lunch. I'll leave the above as it is. I'll finish it tomorrow. It is energy intensive to draw these words out. It seems finishing the picture drained the energy out of me. The guys in the /beg/ thread are recommending Twitter. Twitter it shall be then. I have an old handle there from a decade ago that I never used. I'll dust it off and give it a try after I finish the review.

Tomorrow, I think I'll take a break from doing art. After finishing the review and post it on Twitter, I'll have to think what my next move is. Probably doing more of the limbo and maybe doing some cell shaded art of it. After that I'll have to get back to sculpting. I need to further improve that base mesh, and turn it into actual model with hair and colored eyes. I'll have to retopo and rig it as well. All this will take me a while, but if I can accomplish these goals, I should be just about good enough to do the art parts of Heaven's Key. It will take some work to do just the cover. A floating city covered in a golden radiance will be a decent challenge for me."