"11am. As it turns out, Dohna Dohna got translations two months ago so after finishing the latest Gup OVA that is what I've been playing. It was such an unexpected thing, like a late Chrismas present.
What is on my list today?

https://www.youtube.com/results?search_query=digitial+music+composition+beginner

https://youtu.be/x7uQReA6o8M
Ultimate Beginner's Guide To Music Production

I think I'll start with this.

https://youtu.be/AbyfuW5Aocg
Blender Tutorial - Cloth Simulation | Advance Simulation | Blender 2.9

https://www.youtube.com/c/Iridesium/videos

After that I'll do some of this.

11:30am. Let me start.

Games that taught me that when you win a fight, the failures taken to get to that point stop mattering.

So that is how I will get over my regrets. I've been fighting, but I am yet to win even once. But I know what to do, so I won't act like a madman again. From here on out, that will be the job of the machines.

I've been thinking about genetic programming, and my prior is that the stronger the types, the better the results would be. Mutating an assembly language would be very brittle. I think that DNA code is a functionaly pure, async language. It is certainly nothing like computer assembly.

The success of the future endeavor will depend on how good I am at setting up the environment.

Poker is perfectly enough as a game.

But to go beyond, I might have to start work on a dependently typed language. My personal view is that for humans, non-dependently typed languages like F# and Spiral are the pinnacle of design, but for a completely brainless process it might be different. So this language, unlike Spiral might be specifically for machines.

A lot of work has been done in the type theory domain that could use revisiting in light of new hardware.

11:40am. Maybe some genius will come that will illuminate the secrets of intelligence, but more likely than not, the ML researchers will all fail and I will have to pave my own way to the Singularity.

One thing I am sure is that if making a GP system on the new hardware would be hard, working on a DP language would be downright impossible for an ML researcher. I bet at this level of difficulty that would apply to the guys even at places like Deepmind as well.

PL guys could do it, but they are not interested in ML. So I am one of the rare people in the world who can go down this path.

It would really be a lot of work.

I understand that to bring this to a close, I would have to risk the rest of my lifespan. But even if it took me 30 years, it would be fine.

What I need is a just a little help. I can't do this on a 0$ a month income.

So I won't do ML again until I can resolve that. I do not care from where it will come from. But I won't willingly sell myself into slavery to pursue my path. I already had some of that experience from going to school and I do not want to be coerced into bullshit again.

11:50am. Now let me start. Music, come on!

https://youtu.be/x7uQReA6o8M?t=135

Lol, this is nothing. Though I've filled up 1.7tb on my 3tb hard drive, so I might have to do some deleting if it comes to that.

https://youtu.be/x7uQReA6o8M?t=152

I am a bit weak on memory since I have only 8gb. But thankfully that is the cheapest part of a computer.

https://youtu.be/x7uQReA6o8M?t=278

Take a look at his rig in the background. No way do I want all of that. The way I've been imagining doing audio is just stitching clips together. With that kind of workflow I won't need audio interfaces or whatever. This video is also the first time I've heard about virtual instruments. That sounds interesting.

The way I imagine audio should work is like animation.

https://youtu.be/x7uQReA6o8M?t=358

I am going to have to look into how to make audio using just my computer. Who is going to deal with all this hardware.

https://youtu.be/x7uQReA6o8M?t=468

I suppose I could try this once it is time to get started.

https://youtu.be/x7uQReA6o8M?t=606

Music production is srs bsns.

https://youtu.be/x7uQReA6o8M?t=975

Stitching together samples is how I've been imagining low budget music production to go.

https://youtu.be/x7uQReA6o8M?t=1049
> You could start down the road of music production and make decent stuff without any of these.

12:15pm. https://youtu.be/4cFwUXSZgEs
Best Free DAWs [Free Software to Make Music] (2021)

This is a welcome diversion from the monotony of going through those Blender tutorials.

https://youtu.be/4cFwUXSZgEs?t=429

I wonder if the piano tool is just for the piano, or if any instrument can be used with this?

https://youtu.be/4cFwUXSZgEs?t=507
> If you can use this DAW I would highly recommend it.

He says that Cakewalk is the best one on the list.

https://www.youtube.com/watch?v=3qfa9hGJzoY
Kako koristiti LMMS - Vodič za početnike (BESPLATNO DAW)

The video is in English, but the title and the description are in Croatian. This is the first time I've seen something like this. Is Google experimenting with machine translation?

Nevermind that.

https://youtu.be/AJ1XqtcQjrU
How to start making music with a computer

Let me watch this. After that I will look into how rigid body + cloth interaction should work.

https://youtu.be/AJ1XqtcQjrU?t=143
> MIDI doesn't produce music or audio, it just tells the computer what notes to play at what time. Once you use MIDI, you can use a virtual instrument.

MIDI is what I want.

https://youtu.be/AJ1XqtcQjrU?t=274

This is just like I how I imagined it. When it is time to start I'll just get Cakewalk and dive into it.

https://youtu.be/AJ1XqtcQjrU?t=289
> This is the cool thing about MIDI, it's that since the MIDI itself is not tied to the sound you hear, you can swap out the different sounds, but keep the original melody that you played.

So it is a like painting with gradient maps. I like that.

This is a very good beginner tutorial.

1pm. Enough of music. Though I am looking forward to it, music just on its own is useless for me. I need art and music to support my writing. In isolation, my only choice would be to collaborate with other people, and I'd never get the work that I want that way. It will be more fun this way.

It does not matter if I can only make peanuts like this. I've decided on the price that I want to pay, and I will follow through with it.

https://www.youtube.com/results?search_query=blender+rigid+body+cloth

I am going to have to go through all of these in order to figure it out.

https://youtu.be/AbyfuW5Aocg
Blender Tutorial - Cloth Simulation | Advance Simulation | Blender 2.9

https://www.youtube.com/c/Iridesium/videos

Let me bring these up. I have just a bit left to go and after that I'll be able to move forward into doing my own things.

The basics are what is the most important, but it can be hard to know what the basics are unless you've touched upon everything.

1:05pm. It is really hard to get what you want. Right now I am struggling and trying to barge my way into the AI scene uninvited, but if I gave it time, the incompetence of the rest of the field will roll out the red carpet for me. Right now I want those AI chip companies to sponsor me, but once they ditch their deep learning obsession and start making hardware that has a general purpose, the demnd for the language would come.

Just these Blender simulations would be a great use case for AI chips. Just that cloth simulation grinds the FPS to 1. With dedicated hardware for graph computations, you'd get much better results.

1:10pm. My failures are something that I've invited myself by not anticipating what I would need.

The ML field is incompetent at their job, but there is no need to join them in that. I'll be able to cultivate my path to the pinnacle once I get the pieces that I need. I know that the genetic programming plan will succeed given enough resources.

Until then why not have some fun?"